Use doubles where needed

+++ src/main/java/de/uni_oldenburg/transport/Solution.java
@@ -106,7 +106,7 @@ public class Solution {
 	}
 
 	public double getTotalConsumption() {
-		int consumption = 0;
+		double consumption = 0;
 		for (Tour tour : truckTours) {
 			consumption += tour.getConsumption();
 		}
